feat: add mongodb url to sql checker call
