+++ modules/fbs-core/api/src/main/scala/de/thm/ii/fbs/model/checker/Submission.scala
@@ -1,0 +1,9 @@
+package de.thm.ii.fbs.model.checker
+
+import com.fasterxml.jackson.databind.JsonNode
+
+trait Submission {
+  val id: Int
+  val user: User
+  def toJson: JsonNode
+}
